infra/drive_poller/requirements.txt
google-cloud-pubsub==2.21.1
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
flask==3.0.0

